Mid-work commit
Just working through the last section of the course, building the
classic Todo List.  The course uses a lot of on().  I switched these to
listenTo.  I also have each TodoItem view listening for the destruction
of its model and then removing itself; the course has the ListView
doing the removing.  I think this is a little backwards and that the
todo view should take care of itself and then the list view does have
to care.